+++ src/Sentry/TransactionTracer.cs
@@ -217,12 +217,12 @@ namespace Sentry
             StartChild(SpanId, operation);
 
         /// <inheritdoc />
-        public void Finish(SpanStatus status = SpanStatus.Ok)
+        public void Finish()
         {
             try
             {
+                Status ??= SpanStatus.UnknownError;
                 EndTimestamp = DateTimeOffset.UtcNow;
-                Status = status;
 
                 // Client decides whether to discard this transaction based on sampling
                 _hub.CaptureTransaction(new Transaction(this));
@@ -230,11 +230,15 @@ namespace Sentry
             finally
             {
                 // Clear the transaction from the scope
-                _hub.ConfigureScope(scope =>
-                {
-                    scope.ResetTransaction(this);
-                });
+                _hub.ConfigureScope(scope => scope.ResetTransaction(this));
             }
+        }
+
+        /// <inheritdoc />
+        public void Finish(SpanStatus status)
+        {
+            Status = status;
+            Finish();
         }
 
         /// <inheritdoc />