+++ test/Sentry.Tests/Internals/Http/CachingTransportTests.cs
@@ -56,7 +56,7 @@ namespace Sentry.Tests.Internals.Http
             sentEnvelope.Should().BeEquivalentTo(envelope, o => o.Excluding(x => x.Items[0].Header));
         }
 
-        [Fact(Timeout = 7000)]
+        [Fact(Timeout = 7000, Skip = "Flaky")]
         public async Task EnvelopeReachesInnerTransport()
         {
             // Arrange